Package files accordingly
