+++ .gitignore
@@ -16,5 +16,5 @@ bin/
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
 
-# Mumbo data
+# mumbo.mumbo data
 mumbo_tasks.txt
