+++ src/main/java/mumbo/exception/MumboException.java
@@ -1,3 +1,8 @@
+package mumbo.exception;
+
+/**
+ * A simple runtime exception specifically from Mumbo
+ */
 public class MumboException extends RuntimeException {
     public MumboException(String message) {
         super(message);
