+++ src/main/java/mumbo/task/Event.java
@@ -1,7 +1,11 @@
+package mumbo.task;
+
 import java.time.LocalDateTime;
 
+import mumbo.userinput.DateTimeUtil;
+
 /**
- * Event class
+ * Mumbo.Event class
  *
  * A type of class that has a description, a start and an end
  */
@@ -10,6 +14,12 @@ public class Event extends Task {
     protected LocalDateTime start;
     protected LocalDateTime end;
 
+    /**
+     * Creates an Event task
+     * @param task a String depicting the task's description
+     * @param start a LocalDateTime object that signifies the start of the event
+     * @param end a LocalDateTime object that signifies the end of the event
+     */
     public Event(String task, LocalDateTime start, LocalDateTime end) {
         super(TaskType.EVENT, task);
         this.start = start;
