+++ src/main/java/mumbo/task/TaskType.java
@@ -1,3 +1,5 @@
+package mumbo.task;
+
 public enum TaskType {
     TODO("T"), DEADLINE("D"), EVENT("E");
